Create executorch specific ops file
