+++ torchao/quantization/prototype/qat.py
@@ -16,198 +16,183 @@ from torchao.quantization.unified import TwoStepQuantizer
 
 
 if TORCH_VERSION_AFTER_2_4:
-    from torchao.quantization.GPTQ import (
-        _replace_linear_8da4w,
-        Int8DynActInt4WeightLinear,
-    )
+from torchao.quantization.GPTQ import (
+    _replace_linear_8da4w,
+    Int8DynActInt4WeightLinear,
+)
 
-    class Int8DynActInt4WeightQATQuantizer(TwoStepQuantizer):
-        """
-        Quantizer for performing QAT on a model, where linear layers have int8
-        dynamic per token fake quantized activations and int4 fake quantized
-        grouped per channel weights.
-        """
+class Int8DynActInt4WeightQATQuantizer(TwoStepQuantizer):
+    """
+    Quantizer for performing QAT on a model, where linear layers have int8
+    dynamic per token fake quantized activations and int4 fake quantized
+    grouped per channel weights.
+    """
 
-        def __init__(
-            self,
-            groupsize: int = 256,
-            padding_allowed: bool = False,
-            precision: torch.dtype = torch.float32,
-            scales_precision: torch.dtype = torch.float32,
-        ) -> None:
-            super().__init__()
-            self.groupsize: int = groupsize
-            self.padding_allowed: bool = padding_allowed
-            self.precision: torch.dtype = precision
-            self.scales_precision: torch.dtype = scales_precision
+    def __init__(
+        self,
+        groupsize: int = 256,
+        padding_allowed: bool = False,
+        precision: torch.dtype = torch.float32,
+        scales_precision: torch.dtype = torch.float32,
+    ) -> None:
+        super().__init__()
+        self.groupsize: int = groupsize
+        self.padding_allowed: bool = padding_allowed
+        self.precision: torch.dtype = precision
+        self.scales_precision: torch.dtype = scales_precision
 
-        def prepare(
-            self,
-            model: torch.nn.Module,
-            *args: Any,
-            **kwargs: Any
-        ) -> torch.nn.Module:
-            _replace_linear_8da4w(
-                model,
+    def prepare(
+        self,
+        model: torch.nn.Module,
+        *args: Any,
+        **kwargs: Any
+    ) -> torch.nn.Module:
+        _replace_linear_8da4w(
+            model,
+            self.groupsize,
+            self.padding_allowed,
+            self.precision,
+            self.scales_precision,
+            Int8DynActInt4WeightQATLinear,
+            copy_weights=True,
+        )
+        return model
+
+    def convert(
+        self,
+        model: torch.nn.Module,
+        *args: Any,
+        **kwargs: Any
+    ) -> torch.nn.Module:
+        _convert_qat_linear_8da4w(model)
+        return model
+
+def _convert_qat_linear_8da4w(module: torch.nn.Module):
+    """
+    Replace all `Int8DynActInt4WeightQATLinear` with `Int8DynActInt4WeightLinear`.
+    """
+    for name, child in module.named_children():
+        if isinstance(child, Int8DynActInt4WeightQATLinear):
+            quantized_linear = Int8DynActInt4WeightLinear(
+                child.in_features,
+                child.out_features,
+                bias=False,
+                groupsize=child.groupsize,
+                precision=child.precision,
+                scales_precision=child.scales_precision,
+            )
+            setattr(module, name, quantized_linear)
+
+            # Load weights and qparams into quantized linear
+            n_bit = 4
+            (qmin, qmax) = child._get_qmin_qmax(n_bit)
+            (s, zp) = get_group_qparams_symmetric(child.weight, n_bit, child.groupsize)
+            q_weight = torchao._ops._quantized_decomposed_quantize_per_channel_group(
+                child.weight, s, zp, qmin, qmax, torch.int8, child.groupsize,
+            )
+            quantized_linear.weight = q_weight
+            quantized_linear.scales = s
+            quantized_linear.zeros = zp
+        else:
+            _convert_qat_linear_8da4w(child)
+
+class Int8DynActInt4WeightQATLinear(torch.nn.Linear):
+    """
+    This module implements a linear layer with int8 dynamic per token fake
+    quantized activations with int4 fake quantized grouped per channel weights.
+
+    args:
+        groupsize: the number of elements in each quantized group for weights
+        precision: precision of weights
+        scales_precision: precision of per group scales and zero points
+    """
+
+    def __init__(
+        self,
+        in_features: int,
+        out_features: int,
+        bias: bool = False,
+        device: torch.device = None,
+        groupsize: int = 256,
+        precision: torch.dtype = torch.float32,
+        scales_precision: torch.dtype = torch.float32,
+    ) -> None:
+        super().__init__(
+            in_features,
+            out_features,
+            bias,
+            device=device,
+            dtype=precision,
+        )
+        assert (
+            in_features % groupsize == 0
+        ), f"require in_features:{in_features} % groupsize:{groupsize} == 0"
+        assert not bias, "require bias=False"
+        self.groupsize = groupsize
+        self.precision = precision
+        self.scales_precision = scales_precision
+        # TODO: make this configurable?
+        self.zero_points_precision = torch.int32
+        self._fake_quant_enabled = True
+
+    def enable_fake_quant(self, enabled: bool = True):
+        self._fake_quant_enabled = enabled
+
+    def disable_fake_quant(self):
+        self.enable_fake_quant(False)
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        # activations: int8 dynamic asymmetric quant
+        if self._fake_quant_enabled:
+            (act_scales, act_zp) = _choose_qparams_per_token_asymmetric(
+                x, self.scales_precision, self.zero_points_precision,
+            )
+            (act_qmin, act_qmax) = self._get_qmin_qmax(8)
+            x_fq = fake_quantize_per_token(
+                x, act_scales, act_zp, act_qmin, act_qmax,
+            )
+        else:
+            x_fq = x
+
+        # weights: int4 grouped per channel symmetric quant
+        if self._fake_quant_enabled:
+            (weight_scales, weight_zp) = get_group_qparams_symmetric(
+                self.weight, 4, self.groupsize, self.scales_precision,
+            )
+            # TODO: pass zp dtype to `get_group_qparams_symmetric` instead
+            weight_zp = weight_zp.to(self.zero_points_precision)
+            (weight_qmin, weight_qmax) = self._get_qmin_qmax(4)
+            w_fq = fake_quantize_per_channel_group(
+                self.weight,
+                weight_scales,
+                weight_zp,
+                weight_qmin,
+                weight_qmax,
                 self.groupsize,
-                self.padding_allowed,
-                self.precision,
-                self.scales_precision,
-                Int8DynActInt4WeightQATLinear,
-                copy_weights=True,
             )
-            return model
+        else:
+            w_fq = self.weight
+        return torch.nn.functional.linear(x_fq, w_fq)
 
-        def convert(
-            self,
-            model: torch.nn.Module,
-            *args: Any,
-            **kwargs: Any
-        ) -> torch.nn.Module:
-            _convert_qat_linear_8da4w(model)
-            return model
+    # TODO: move this to common util
+    def _get_qmin_qmax(self, n_bit: int):
+        qmin = -(2 ** (n_bit - 1))
+        qmax = 2 ** (n_bit - 1) - 1
+        return (qmin, qmax)
 
-    def _convert_qat_linear_8da4w(module: torch.nn.Module):
-        """
-        Replace all `Int8DynActInt4WeightQATLinear` with `Int8DynActInt4WeightLinear`.
-        """
-        for name, child in module.named_children():
-            if isinstance(child, Int8DynActInt4WeightQATLinear):
-                quantized_linear = Int8DynActInt4WeightLinear(
-                    child.in_features,
-                    child.out_features,
-                    bias=False,
-                    groupsize=child.groupsize,
-                    precision=child.precision,
-                    scales_precision=child.scales_precision,
-                )
-                setattr(module, name, quantized_linear)
+def enable_8da4w_fake_quant(mod: torch.nn.Module):
+    """
+    Enable fake quantization for `Int8DynActInt4WeightQATLinear`.
+    """
+    if isinstance(mod, Int8DynActInt4WeightQATLinear):
+        mod.enable_fake_quant()
 
-                # Load weights and qparams into quantized linear
-                n_bit = 4
-                (qmin, qmax) = child._get_qmin_qmax(n_bit)
-                (s, zp) = get_group_qparams_symmetric(child.weight, n_bit, child.groupsize)
-                q_weight = torch.ops.quantized_decomposed.quantize_per_channel_group(
-                    child.weight, s, zp, qmin, qmax, torch.int8, child.groupsize,
-                )
-                quantized_linear.weight = q_weight
-                quantized_linear.scales = s
-                quantized_linear.zeros = zp
-            else:
-                _convert_qat_linear_8da4w(child)
-
-    class Int8DynActInt4WeightQATLinear(torch.nn.Linear):
-        """
-        This module implements a linear layer with int8 dynamic per token fake
-        quantized activations with int4 fake quantized grouped per channel weights.
-
-        args:
-            groupsize: the number of elements in each quantized group for weights
-            precision: precision of weights
-            scales_precision: precision of per group scales and zero points
-        """
-
-        def __init__(
-            self,
-            in_features: int,
-            out_features: int,
-            bias: bool = False,
-            device: torch.device = None,
-            groupsize: int = 256,
-            precision: torch.dtype = torch.float32,
-            scales_precision: torch.dtype = torch.float32,
-        ) -> None:
-            super().__init__(
-                in_features,
-                out_features,
-                bias,
-                device=device,
-                dtype=precision,
-            )
-            assert (
-                in_features % groupsize == 0
-            ), f"require in_features:{in_features} % groupsize:{groupsize} == 0"
-            assert not bias, "require bias=False"
-            self.groupsize = groupsize
-            self.precision = precision
-            self.scales_precision = scales_precision
-            # TODO: make this configurable?
-            self.zero_points_precision = torch.int32
-            self._fake_quant_enabled = True
-
-        def enable_fake_quant(self, enabled: bool = True):
-            self._fake_quant_enabled = enabled
-
-        def disable_fake_quant(self):
-            self.enable_fake_quant(False)
-
-        def forward(self, x: torch.Tensor) -> torch.Tensor:
-            # activations: int8 dynamic asymmetric quant
-            if self._fake_quant_enabled:
-                (act_scales, act_zp) = _choose_qparams_per_token_asymmetric(
-                    x, self.scales_precision, self.zero_points_precision,
-                )
-                (act_qmin, act_qmax) = self._get_qmin_qmax(8)
-                x_fq = fake_quantize_per_token(
-                    x, act_scales, act_zp, act_qmin, act_qmax,
-                )
-            else:
-                x_fq = x
-
-            # weights: int4 grouped per channel symmetric quant
-            if self._fake_quant_enabled:
-                (weight_scales, weight_zp) = get_group_qparams_symmetric(
-                    self.weight, 4, self.groupsize, self.scales_precision,
-                )
-                # TODO: pass zp dtype to `get_group_qparams_symmetric` instead
-                weight_zp = weight_zp.to(self.zero_points_precision)
-                (weight_qmin, weight_qmax) = self._get_qmin_qmax(4)
-                w_fq = fake_quantize_per_channel_group(
-                    self.weight,
-                    weight_scales,
-                    weight_zp,
-                    weight_qmin,
-                    weight_qmax,
-                    self.groupsize,
-                )
-            else:
-                w_fq = self.weight
-            return torch.nn.functional.linear(x_fq, w_fq)
-
-        # TODO: move this to common util
-        def _get_qmin_qmax(self, n_bit: int):
-            qmin = -(2 ** (n_bit - 1))
-            qmax = 2 ** (n_bit - 1) - 1
-            return (qmin, qmax)
-
-    def enable_8da4w_fake_quant(mod: torch.nn.Module):
-        """
-        Enable fake quantization for `Int8DynActInt4WeightQATLinear`.
-        """
-        if isinstance(mod, Int8DynActInt4WeightQATLinear):
-            mod.enable_fake_quant()
-
-    def disable_8da4w_fake_quant(mod: torch.nn.Module):
-        """
-        Disable fake quantization for `Int8DynActInt4WeightQATLinear`.
-        """
-        if isinstance(mod, Int8DynActInt4WeightQATLinear):
-            mod.disable_fake_quant()
-
-else:  # not TORCH_VERSION_AFTER_2_4
-
-    class Int8DynActInt4WeightQATQuantizer:
-        def __init__(*args, **kwargs):
-            raise ValueError(
-                "Int8DynActInt4WeightQATQuantizer is only supported after PyTorch 2.4+"
-            )
-
-    class Int8DynActInt4WeightQATLinear:
-        def __init__(*args, **kwargs):
-            raise ValueError(
-                "Int8DynActInt4WeightQATLinear is only supported after PyTorch 2.4+"
-            )
-
+def disable_8da4w_fake_quant(mod: torch.nn.Module):
+    """
+    Disable fake quantization for `Int8DynActInt4WeightQATLinear`.
+    """
+    if isinstance(mod, Int8DynActInt4WeightQATLinear):
+        mod.disable_fake_quant()
 
 # ========================
 # |   QUANT PRIMITIVES   |
